Add Stella Strategy (#890)
* add Stella Strategy

* hide feeBps if null

* fix as requested

* fix feeBps calc

* update baseApy to wide range
